refactor(RestResponse.ts) : move directory
