+++ src/apis/auth/hooks/useAuthLoginMutation.ts
@@ -1,8 +1,7 @@
 import { useMutation } from "@tanstack/react-query";
 
 import { AuthLoginRequest } from "@apis/auth/models/AuthLogin";
-
-import { RestErrorResponse, RestResponse } from "@models/RestResponse";
+import { RestErrorResponse, RestResponse } from "@apis/_models/RestResponse";
 
 import rest from "@utils/rest";
